requirements.txt
colorthief==0.2.1
disnake==2.4.0
Pillow==9.1.0
pyowo==1.0.1
qrcode==7.3.1
requests==2.27.1